requirements.txt
customtkinter>=5.2.0
llama-cpp-python>=0.2.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
cryptography>=41.0.0
pillow>=10.0.0
psutil>=5.9.0
configparser>=5.3.0
